Rename module to invariant_sdk and add pyproject.toml instead of requirements.txt
