--- python/requirements.txt
@@ -1,4 +1,0 @@
-pydantic
-pytest
-pytest-cov
-requests
